+++ src/main/java/com/scaef/scaef_backend/repository/MedicamentoRepository.java
@@ -6,8 +6,8 @@ package com.scaef.scaef_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface MedicamentoRepository extends JpaRepository <Medicamento, Integer> {
+public interface MedicamentoRepository extends JpaRepository <Medicamento, Long> {
     @Query(value = "select id from Medicamento where id =:id", nativeQuery = true)
-    boolean exist(int id);
+    boolean exist(Long id);
     
 }
