+++ src/main/java/com/scaef/scaef_backend/service/MedicamentoService.java
@@ -28,7 +28,7 @@ public class MedicamentoService {
         medicamentoRepository.save(medicamento);
     }
 
-    public Optional<Medicamento> findById(int id) {
+    public Optional<Medicamento> findById(Long id) {
         return medicamentoRepository.findById(id);
     }
